+++ @xen-orchestra/web/src/components/host/dashboard/HostDashboardCpuProvisioning.vue
@@ -16,7 +16,6 @@
 import { useHostStore } from '@/stores/xo-rest-api/host.store'
 import { useVmStore } from '@/stores/xo-rest-api/vm.store'
 import type { XoHost } from '@/types/xo/host.type'
-import { VM_POWER_STATE } from '@/types/xo/vm.type.ts'
 import VtsLoadingHero from '@core/components/state-hero/VtsLoadingHero.vue'
 import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiCardNumbers from '@core/components/ui/card-numbers/UiCardNumbers.vue'
@@ -42,17 +41,7 @@ const hostVms = computed(() => vmsByHost.value.get(host.id) ?? [])
 
 const cpusCount = computed(() => host.cpus.cores)
 
-const vCpusCount = useArrayReduce(
-  hostVms,
-  (total, vm) => {
-    if (vm.power_state !== VM_POWER_STATE.RUNNING && vm.power_state !== VM_POWER_STATE.PAUSED) {
-      return total
-    }
-
-    return total + vm.CPUs.number
-  },
-  0
-)
+const vCpusCount = useArrayReduce(hostVms, (total, vm) => total + vm.CPUs.number, 0)
 </script>
 
 <style lang="postcss" scoped>